Previous posts can now be seen offline in the FeedFragment and the ProfileFragment. Profile can be edited, and the change will be pinned and then saved when the user connects to internet again.
User can not add a new post because the ParseFile doesn’t yet exist because it has not been saved to Parse. The user can pin it but it will not be added to the RV to view until it is added to Back4App. To avoid an error and combat this for now, I added a conditional that checks for internet connection: if the user has the connection then save post and query else toast that the user needs to connect to internet to continue.

Minor Changes: User can change Username or Bio without editing both of them.